refactor:#104 파일명 및 컴포넌트명 수정 / 메인 페이지에 컴포넌트 적용
