+++ src/components/features/home/main-random-mission.tsx
@@ -2,7 +2,7 @@ import { getUserSessionState } from '@/lib/utils/api/auth-action';
 import ImageSection from '@/components/features/home/main-random-mission-image';
 import MissionButton from '@/components/features/home/main-random-mission-button';
 import RandomMissionContainer from '@/components/features/home/main-random-mission-container';
-import TitleSection from '@/components/features/home/main-random-title-box';
+import TitleSection from '@/components/features/home/main-random-title';
 
 const RandomMissionSection = async () => {
   const { isLogin } = await getUserSessionState();
